Server side M2M4Ria.M2MEntityCollection is now generated within EntityModelNamespace.
This means there is one M2M4Ria.M2MEntityCollection implementation generated
per entity model, but each is generated in a unique name space.
    
The reason to generate M2M4Ria.M2MEntityCollection and not having it in a
separate .cs file is ease of deployment. For M2M4Ria we only need to deploy
a couple of .tt and .ttinclude files, and not also a separate dll's.