Changed the implementation of removing link table entities.
Previously there was an ..Removed action for each entity type
that was contained in an M2M relation. This action was initialized
by the domain context when an entity of that type was added to a
corresponding entity set (and reset to null when the entity was removed).
    
In the new situation, each join table type contains a property EntitySet
that gives access to the entity's entity set (RIA should have provided
this property on every entity, ... may be next version ...). This property
is set/reset by the domain context when an entity is added/removed to/from
the corresponding entity set.
    
The remove actions of the entities that are part of an M2M relation now
call the Remove method on either their own entity collection or on the
entity set of the join table entry that is to be removed. Which Remove
action is invoked depends on whether the join table entity is present in
an entity set or not.
    
This new approach is a little more cleaner and requires little modification
when MS decides to add the ability to entities to access their entity sets.
Moreover, since the EntitySet property is on join table entities, fewer
handlers need to be set in the domain context, fewer event handlers will be
called, and less memory is occupied for administering entity sets.