M2M collection of client-side entities are now of type IEntityCollection<TEntity>
This used to be EntityCollection<JoinType,TEntity>, but this was overly
complicated and unnecessarily exposed the relation between TEntity and
JoinType. The type shown by Intellisense was not easy to understand.
    
To hide the JoinType information from the collection type, a new interface
IEntityCollection<TEntity> is used. Internally the type
EntityCollection<JoinType,TEntity> is still used.