Optionally generate public IQueryable<JointType> Get<JoinType>Set() domain service operation
This feature is added to support a more incremental introduction of m2m4ria
in your projects.
    
Generation is controlled with the new CreateQueryOperationForJoinType configuration switch
    
If set to 'true' RIA generates an entity set in the domain context for the join type (e.g., ctx.DogTrainers) that you can use for querying, inserting and deleting join type objects.
    
This enables you to incrementally replace the use of join type objects in your projects with using m2m relations.