Replaced CreateQueryOperationForJoinType property by IsJoinTypeEntitySetAccessibleFromDomainContext that only affects the generated domain context
The IsJoinTypeEntitySetAccessibleFromDomainContext property adds a property to the domain context that gives access to the entityset of join type objects.
    
This is to simplify incremental adoption of m2m4ria into your projects because you now still have access to join type entity sets (e.g., ctx.DogTrainers) to add/remove joint type objects.
    
The current approach only adds a property to the domain context, the approach with the CreateQueryOperationForJoinType property also generated a query operation in the domain service, while that is actually not needed.