Bug fix: incorrect NotifyCollectionChanged event was generated in client
The NotifyCollectionChangedEventArgs always used 0 (zero) as index.
As result, if a collection was bound in Silverlight to e.g., a ListBox 
control, deleting an item always resulted in removing the first item from
the ListBox. 
Note that this bug didn't affect the database. In the database always the 
correct item was removed.