M2M4Ria: Improved implementation of client-side EntityCollection.Add method.
When calling the Add method a new join type entity is created. Whenever
one of the navigation properties is set, an INotifyCollectionChanged event
is fired for the join type collection. Previously, the second navigation 
property of the newly created join type element was still null. This has now
been fixed: both navigation properties will have a value when an 
INotifyCollectionChanged event is fired.