m2m4ria: Bug fix: deleting an m2m relation for an entity that is also deleted didn't work
This scenario is demonstrated in DeleteParentTest test in DomainContextTest.cs

1) Create a Dog and a Vet objects and a DogVet relation
2) Save these to the database
3) Remove (delete) the DogVet relation and the Dog object from the database
This resulted in an error, because in the DeleteAnimalVet domain service 
operation the vet object is attached to the dog object which can't be 
done if dog has an entity state of EntityState.Deleted. 

Fix. In case dog has a Deleted entity state, it is temporarily set to 
EntityState.Unchanged. Then vet is attached to dog, and finally 
dog.EntityState is reset to EntityState.Deleted.