M2M4RIA: Fixed bug as reported in issuw 15877
The problem was that names of association ends did not correspond to entity type names when creating an EF model from an existing data base _and_ the pluralization engine of EF was used.

Fix. Don't depend on association end names but call GetEntityType on each end to retrieve the correct entity types.