M2M4RIA: Upgraded to RIA services version 1.SP1
The most significant change in the latest release of RIA services is that
an entity's entity set can now be accessed. This requires using
the IExtendedEntity interface from http://riaservicescontrib.codeplex.com.

This new feature of RIA services simplifies the client-side code that is 
generated by m2m4ria.

NOTE. If you're not using SP1, don't use this version of m2m4ria.