Bug fix: deleting m2m association resulted in discarding of other changes (issues 16713, 16714, 16715)
The problem was in the LoadM2M extension method where Detach/Attach was used to
delete an existing m2m association. However, as a result of using detach, all remaining
pending changes of the detached entity where discarded.

Fix
Now using the m2m association from the database. This is not the most optimal solution
but see http://stackoverflow.com/questions/13417463/entityframework-remove-m2m-association-given-only-primary-keys-how
and http://stackoverflow.com/questions/13528531/entityframework-possible-many-2-many-bug-in-ef-5-0